Add a modified version of the OSD dataset
Object Segmentation Database is a popular dataset for evaluation of
object segmentation algorithms. This is a modified version, published
with permission from the author of the original dataset. The main change
is improvement of labels near object boundaries and introduction of more
flexible labeling scheme that allows this dataset to be used for
evaluation of both model-based and model-free segmentation algorithms.
See accompanying README.md for more details.